Move functions that may be needed for multiple test files to utils.py #27
